[runtime] move rust runtime to lib and cmd
