--- .github/workflows/examples.yml
@@ -1,32 +1,0 @@
-name: examples
-
-on:
-  pull_request:
-    branches: "*"
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: get submodule
-      run:  git submodule update --init --recursive
-    - name: install deps
-      run: |
-          sudo make install-deps
-          make /opt/wasi-sdk
-          wget https://github.com/tinygo-org/tinygo/releases/download/v0.27.0/tinygo_0.27.0_amd64.deb
-          sudo dpkg -i tinygo_0.27.0_amd64.deb
-    - name: make
-      run: make -C runtime/cpp
-    - name: Install rust stable
-      uses: actions-rs/toolchain@v1
-      with:
-          toolchain: stable
-          override: true
-    - name: examples with rust runtime
-      run: IMPL=rust make -C examples
-    - name: examples with cpp runtime
-      run: IMPL=cpp  make -C examples
