+++ runtime/Cargo.toml
@@ -1,0 +1,10 @@
+[workspace]
+members = [
+    "cmd", "wasm-bpf-rs"
+]
+
+[profile.release]
+opt-level = 3
+lto = true
+debug = 0
+overflow-checks = false
