+++ runtime/cmd/.gitignore
@@ -1,0 +1,4 @@
+target
+dump*
+*.wasm
+wasm-bpf
