+++ runtime/cmd/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "wasm-bpf"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+description = "A WebAssembly runtime cmd for eBPF programs based on libbpf and wasmtime (part of wasm-bpf project)"
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.69"
+clap = { version = "4.1.4", features = ["derive"] }
+flexi_logger = "0.25.1"
+log = "0.4.17"
+wasm-bpf-rs = { path = "../wasm-bpf-rs" }
