+++ runtime/cmd/Makefile
@@ -1,0 +1,4 @@
+build:
+	cargo build --release
+clean:
+	cargo clean
