--- runtime/rust/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "wasm-bpf-rs"
-version = "0.1.0"
-edition = "2021"
-license = "MIT"
-description = "A WebAssembly runtime for eBPF programs based on libbpf and wasmtime (part of wasm-bpf project)"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.69"
-clap = { version = "4.1.4", features = ["derive"] }
-flexi_logger = "0.25.1"
-libbpf-rs = "0.19.1"
-log = "0.4.17"
-wasmtime = "5.0.0"
-wasmtime-wasi = "5.0.0"
-
-[profile.release]
-opt-level = 3
-lto = true
-debug = 0
-overflow-checks = false
