--- runtime/rust/Makefile
@@ -1,4 +1,0 @@
-build:
-	cargo build --release
-clean:
-	cargo clean
