--- runtime/rust/src/log_format.rs
@@ -1,22 +1,0 @@
-//!  SPDX-License-Identifier: MIT
-//!
-//! Copyright (c) 2023, eunomia-bpf
-//! All rights reserved.
-//!
-use flexi_logger::{DeferredNow, Record, TS_DASHES_BLANK_COLONS_DOT_BLANK};
-
-pub fn my_log_format(
-    w: &mut dyn std::io::Write,
-    now: &mut DeferredNow,
-    record: &Record,
-) -> Result<(), std::io::Error> {
-    write!(
-        w,
-        "[{}] {} [{}:{}] {}",
-        now.format(TS_DASHES_BLANK_COLONS_DOT_BLANK),
-        record.level(),
-        record.module_path().unwrap_or("<unnamed>"),
-        record.line().unwrap_or(0),
-        &record.args()
-    )
-}
