+++ runtime/wasm-bpf-rs/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "wasm-bpf-rs"
+version = "0.2.0"
+edition = "2021"
+license = "MIT"
+description = "A WebAssembly runtime cmd for eBPF programs based on libbpf and wasmtime (part of wasm-bpf project)"
+
+[dependencies]
+anyhow = "1.0.69"
+flexi_logger = "0.25.1"
+libbpf-rs = "0.19.1"
+log = "0.4.17"
+wasmtime = "5.0.0"
+wasmtime-wasi = "5.0.0"
+
